+++ WisdomOfAzeroth/WisdomOfAzeroth.lua
@@ -1,48 +1,71 @@
-local f = CreateFrame("Frame", "WisdomOfAzerothFrame", UIParent, "BasicFrameTemplateWithInset")
-f:SetPoint("CENTER")
+Wisdom = {}
+Wisdom.currentIndex = 1
 
--- Fond moderne
-local bg = f:CreateTexture(nil, "BORDER")
+-- Création du frame
+Wisdom.f = CreateFrame("Frame", "WisdomOfAzerothFrame", UIParent, "BasicFrameTemplateWithInset")
+Wisdom.f:SetPoint("CENTER")
+Wisdom.f:SetWidth(350)
+
+-- Fond
+local bg = Wisdom.f:CreateTexture(nil, "BORDER")
 bg:SetAllPoints(true)
 bg:SetAlpha(0.9)
 
 -- Titre
-f.title = f:CreateFontString(nil, "OVERLAY", "GameFontHighlight")
-f.title:SetPoint("TOP", 0, -5)
-f.title:SetText("Wisdom of Azeroth")
+Wisdom.title = Wisdom.f:CreateFontString(nil, "OVERLAY", "GameFontHighlight")
+Wisdom.title:SetPoint("TOP", 0, -5)
+Wisdom.title:SetText("Wisdom of Azeroth")
 
--- Numéro du proverbe
-local index = math.random(#WOW_QUOTES)
-local n = f:CreateFontString(nil, "OVERLAY", "GameFontNormal")
-n:SetPoint("TOP", f.title, "BOTTOM", 0, -10)
-n:SetText("Proverbe n°" .. index .. "/" .. #WOW_QUOTES)
+-- Numéro
+Wisdom.n = Wisdom.f:CreateFontString(nil, "OVERLAY", "GameFontNormal")
+Wisdom.n:SetPoint("TOP", Wisdom.title, "BOTTOM", 0, -10)
 
--- Texte du proverbe
-local t = f:CreateFontString(nil, "OVERLAY", "GameFontNormalLarge")
-t:SetPoint("TOP", n, "BOTTOM", 0, -15)
-t:SetWidth(300)              -- largeur max du texte
-t:SetWordWrap(true)          -- active le retour à la ligne
-t:SetJustifyH("CENTER")      -- centrage horizontal
-t:SetJustifyV("TOP")         -- alignement vertical
-t:SetText(WOW_QUOTES[index])
+-- Texte
+Wisdom.t = Wisdom.f:CreateFontString(nil, "OVERLAY", "GameFontNormalLarge")
+Wisdom.t:SetPoint("TOP", Wisdom.n, "BOTTOM", 0, -15)
+Wisdom.t:SetWidth(300)
+Wisdom.t:SetWordWrap(true)
+Wisdom.t:SetJustifyH("CENTER")
+Wisdom.t:SetJustifyV("TOP")
 
--- Ajustement automatique de la hauteur du cadre
-local neededHeight = t:GetStringHeight() + 115
-f:SetWidth(350)
-f:SetHeight(neededHeight)
+-- Fonction d'affichage
+function Wisdom.ShowQuote(index)
+    Wisdom.currentIndex = index
 
--- bouton suivant
------------------
+    if Wisdom.currentIndex < 1 then
+        Wisdom.currentIndex = #WOW_QUOTES
+    elseif Wisdom.currentIndex > #WOW_QUOTES then
+        Wisdom.currentIndex = 1
+    end
 
-local boutonNext = CreateFrame("Button", nil, f, "UIPanelButtonTemplate")
-boutonNext:SetSize(68, 25)
-boutonNext:SetText("Suivant")
-boutonNext:SetPoint("BOTTOM", f, "BOTTOM", 40, 10)
+    Wisdom.n:SetText("Proverbe n°" .. Wisdom.currentIndex .. "/" .. #WOW_QUOTES)
+    Wisdom.t:SetText(WOW_QUOTES[Wisdom.currentIndex])
 
--- bouton précédent
--------------------
+    -- local neededHeight = Wisdom.t:GetStringHeight() + 115
+    -- Wisdom.f:SetHeight(neededHeight)
 
-local boutonPrevious = CreateFrame("Button", nil, f, "UIPanelButtonTemplate")
-boutonPrevious:SetSize(68, 25)
+    Wisdom.f:SetHeight(160)
+    Wisdom.f:Show()
+end
+
+-- Premier affichage
+Wisdom.ShowQuote(math.random(#WOW_QUOTES))
+
+-- Bouton Previous
+local boutonPrevious = CreateFrame("Button", nil, Wisdom.f, "UIPanelButtonTemplate")
+boutonPrevious:SetSize(75, 25)
 boutonPrevious:SetText("Précédent")
-boutonPrevious:SetPoint("BOTTOM", f, "BOTTOM", -40, 10)
+boutonPrevious:SetPoint("BOTTOM", Wisdom.f, "BOTTOM", -40, 10)
+boutonPrevious:SetScript("OnClick", function()
+    Wisdom.ShowQuote(Wisdom.currentIndex - 1)
+end)
+
+-- Bouton Next
+local boutonNext = CreateFrame("Button", nil, Wisdom.f, "UIPanelButtonTemplate")
+boutonNext:SetSize(75, 25)
+boutonNext:SetText("Suivant")
+boutonNext:SetPoint("BOTTOM", Wisdom.f, "BOTTOM", 40, 10)
+boutonNext:SetScript("OnClick", function()
+    Wisdom.ShowQuote(Wisdom.currentIndex + 1)
+end)
+